ViewModels y Validaciones de CRUD Productos

+++ AppMVC/NicoApp/Models/Producto.cs
@@ -4,13 +4,13 @@ public class Productos
 {
     private int idProducto;
     private string descripcion;
-    private int precio;
+    private decimal precio;
 
     public Productos()
     {
 
     }
-    public Productos(int idProducto, string descripcion, int precio)
+    public Productos(int idProducto, string descripcion, decimal precio)
     {
         this.IdProducto = idProducto;
         this.Descripcion = descripcion;
@@ -19,5 +19,5 @@ public class Productos
 
     public int IdProducto { get => idProducto; set => idProducto = value; }
     public string Descripcion { get => descripcion; set => descripcion = value; }
-    public int Precio { get => precio; set => precio = value; }
+    public decimal Precio { get => precio; set => precio = value; }
 }
